requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
requests==2.32.5
openai-whisper==20250625
pyttsx3==2.99
